Keep file explorer insertion below the Swift length budget
The context-menu insertion implementation now lives in a focused helper file while FileExplorerView keeps only the menu wiring and existing view ownership.

Constraint: CI enforces per-file Swift length budgets.
Rejected: Raising the file length budget | this feature can stay below budget by splitting helper behavior out of the view file.
Confidence: high
Scope-risk: narrow
Directive: Keep path insertion helpers outside FileExplorerView when adding related behavior.
Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv; git diff --check.
Not-tested: Local xcodebuild/tests per repository policy.
