+++ Sources/FileExplorerTerminalPathInsertion.swift
@@ -1,0 +1,144 @@
+import AppKit
+
+enum FileExplorerTerminalPathInsertion {
+    static func insertedText(forPaths paths: [String]) -> String {
+        TerminalImageTransferPlanner.insertedText(forPathStrings: paths)
+    }
+
+    static func insertedText(forPaths paths: [String], relativeToRootPath rootPath: String) -> String {
+        insertedText(forPaths: paths.map { relativePath(for: $0, rootPath: rootPath) })
+    }
+
+    static func relativePath(for path: String, rootPath: String) -> String {
+        var relativePath = path
+        if relativePath.hasPrefix(rootPath) {
+            relativePath = String(relativePath.dropFirst(rootPath.count))
+            if relativePath.hasPrefix("/") {
+                relativePath = String(relativePath.dropFirst())
+            }
+        }
+        return relativePath
+    }
+
+    @discardableResult
+    static func insert(paths: [String], relativeToRootPath rootPath: String? = nil, intoTerminalFor window: NSWindow?) -> Bool {
+        let text: String
+        if let rootPath {
+            text = insertedText(forPaths: paths, relativeToRootPath: rootPath)
+        } else {
+            text = insertedText(forPaths: paths)
+        }
+        guard !text.isEmpty else { return false }
+
+        return MainActor.assumeIsolated {
+            guard let terminalPanel = targetTerminalPanel(for: window) else { return false }
+            terminalPanel.sendText(text)
+            return true
+        }
+    }
+
+    private static func targetTerminalPanel(for window: NSWindow?) -> TerminalPanel? {
+        guard let appDelegate = AppDelegate.shared else { return nil }
+        if let window,
+           let terminalPanel = appDelegate.contextForMainTerminalWindow(window)?.tabManager.selectedWorkspace?.focusedTerminalPanel {
+            return terminalPanel
+        }
+        if let window,
+           let windowId = appDelegate.mainWindowId(from: window),
+           let terminalPanel = appDelegate.tabManagerFor(windowId: windowId)?.selectedWorkspace?.focusedTerminalPanel {
+            return terminalPanel
+        }
+        return appDelegate.tabManager?.selectedWorkspace?.focusedTerminalPanel
+    }
+}
+
+extension NSMenu {
+    func addFileExplorerInsertPathItems(
+        target: AnyObject,
+        representedObject: Any,
+        insertAction: Selector,
+        insertRelativeAction: Selector
+    ) {
+        let insertPathItem = NSMenuItem(
+            title: String(localized: "fileExplorer.contextMenu.insertPath", defaultValue: "Insert Path"),
+            action: insertAction,
+            keyEquivalent: ""
+        )
+        insertPathItem.target = target
+        insertPathItem.representedObject = representedObject
+        addItem(insertPathItem)
+
+        let insertRelativePathItem = NSMenuItem(
+            title: String(localized: "fileExplorer.contextMenu.insertRelativePath", defaultValue: "Insert Relative Path"),
+            action: insertRelativeAction,
+            keyEquivalent: ""
+        )
+        insertRelativePathItem.target = target
+        insertRelativePathItem.representedObject = representedObject
+        addItem(insertRelativePathItem)
+    }
+}
+
+extension FileExplorerPanelView.Coordinator {
+    private func contextMenuNodes(clicked node: FileExplorerNode) -> [FileExplorerNode] {
+        guard let outlineView else { return [node] }
+        let clickedRow = outlineView.clickedRow
+        let selectedRows = outlineView.selectedRowIndexes
+        guard clickedRow >= 0, selectedRows.contains(clickedRow) else {
+            return [node]
+        }
+        let nodes = selectedRows.compactMap { row -> FileExplorerNode? in
+            guard row >= 0, row < outlineView.numberOfRows else { return nil }
+            return outlineView.item(atRow: row) as? FileExplorerNode
+        }
+        return nodes.isEmpty ? [node] : nodes
+    }
+
+    @objc func contextMenuInsertPath(_ sender: NSMenuItem) {
+        guard let node = sender.representedObject as? FileExplorerNode else { return }
+        FileExplorerTerminalPathInsertion.insert(
+            paths: contextMenuNodes(clicked: node).map(\.path),
+            intoTerminalFor: outlineView?.window ?? containerView?.window
+        )
+    }
+
+    @objc func contextMenuInsertRelativePath(_ sender: NSMenuItem) {
+        guard let node = sender.representedObject as? FileExplorerNode else { return }
+        FileExplorerTerminalPathInsertion.insert(
+            paths: contextMenuNodes(clicked: node).map(\.path),
+            relativeToRootPath: store.rootPath,
+            intoTerminalFor: outlineView?.window ?? containerView?.window
+        )
+    }
+}
+
+extension FileExplorerContainerView {
+    private func searchResultsForContextMenu(row: Int) -> [FileSearchResult] {
+        guard row >= 0, row < searchSnapshot.results.count else { return [] }
+        let selectedRows = searchResultsView.selectedRowIndexes
+        guard selectedRows.contains(row) else {
+            return [searchSnapshot.results[row]]
+        }
+        let results = selectedRows.compactMap { selectedRow -> FileSearchResult? in
+            guard selectedRow >= 0, selectedRow < searchSnapshot.results.count else { return nil }
+            return searchSnapshot.results[selectedRow]
+        }
+        return results.isEmpty ? [searchSnapshot.results[row]] : results
+    }
+
+    @objc func contextMenuInsertSearchResultPath(_ sender: NSMenuItem) {
+        guard let row = (sender.representedObject as? NSNumber)?.intValue else { return }
+        FileExplorerTerminalPathInsertion.insert(
+            paths: searchResultsForContextMenu(row: row).map(\.path),
+            intoTerminalFor: window
+        )
+    }
+
+    @objc func contextMenuInsertSearchResultRelativePath(_ sender: NSMenuItem) {
+        guard let row = (sender.representedObject as? NSNumber)?.intValue else { return }
+        FileExplorerTerminalPathInsertion.insert(
+            paths: searchResultsForContextMenu(row: row).map(\.relativePath),
+            intoTerminalFor: window
+        )
+    }
+}
